backend/requirements.txt
flask
flask-cors
cloudinary
pillow
requests
